Consolidate memory clients
