+++ docs/Implementation.md
@@ -418,14 +418,14 @@ print(summary)
 
 - `src/hierarchical_memory.py` now includes optional gRPC support. `MemoryServer`
   wraps a `HierarchicalMemory` instance and serves ``Push`` and ``Query`` RPCs.
-- Helper functions `push_remote()` and `query_remote()` in
-  `src/remote_memory.py` send vectors to the server and retrieve nearest
+  - Helper functions `push_remote()` and `query_remote()` in
+    `src/memory_clients.py` send vectors to the server and retrieve nearest
   neighbours over the network. Asynchronous variants `push_remote_async()` and
   `query_remote_async()` allow non-blocking interaction when using ``grpc.aio``.
 - The server constructor accepts an ``address`` and ``max_workers`` to control
   the bind host and connection pool size.
-- `src/remote_memory.py` provides a small :class:`RemoteMemory` client that wraps
-  these RPCs in a convenient Python interface.
+  - `src/memory_clients.py` provides a small :class:`RemoteMemoryClient` that wraps
+    these RPCs in a convenient Python interface.
 
 ## C-9 Hopfield Associative Memory
 
@@ -709,8 +709,8 @@ train_world_model(model, trajectory_ds, event_dataset=events,
 - Integrate `QAEHyperparamSearch` into `MetaRLRefactorAgent` to tune the exploration rate during refactoring. **Implemented**
 - Rewrite `download_triples()` with asyncio to fetch dataset files
   concurrently. **Implemented** with an async helper using `aiohttp`.
-- Add streaming RPCs to `MemoryServer` so batches of vectors can be pushed and
-  queried in one call. Update `memory.proto` and the `RemoteMemory` client.
+  - Add streaming RPCs to `MemoryServer` so batches of vectors can be pushed and
+    queried in one call. Update `memory.proto` and the `RemoteMemoryClient`.
   **Implemented** via `push_batch_remote()` and `query_batch_remote()` in
   `src/hierarchical_memory.py`.
 - Implement optional gradient checkpointing in `multimodal_world_model.py` via a
@@ -741,7 +741,7 @@ python scripts/attention_analysis.py --model model.pt --input sample.txt --out-d
 - Implement a `DistributedTrainer` that automatically restarts failed
   processes and coordinates checkpoints with `DistributedMemory`. **Implemented**
   in `src/distributed_trainer.py` with tests.
-- Build an `EdgeMemoryClient` to stream context vectors to `RemoteMemory`
+  - Build an `EdgeMemoryClient` to stream context vectors to `RemoteMemoryClient`
   so edge devices can handle large-context inference. **Implemented**
 - The client now keeps a local queue when the network is unreachable and
   periodically flushes queued `add`/`delete` operations once connectivity
