+++ src/distributed_memory.py
@@ -7,7 +7,7 @@ from typing import Iterable, Any, Tuple, List
 import torch
 
 from .hierarchical_memory import HierarchicalMemory
-from .remote_memory import push_remote, query_remote
+from .memory_clients import push_remote, query_remote
 
 
 class DistributedMemory(HierarchicalMemory):
