+++ tests/test_federated_memory_server_proof.py
@@ -3,7 +3,7 @@ import unittest
 import torch
 
 from asi.hierarchical_memory import HierarchicalMemory
-from asi.remote_memory import push_remote, query_remote
+from asi.memory_clients import push_remote, query_remote
 from asi.federated_memory_server import FederatedMemoryServer
 from asi import memory_pb2, memory_pb2_grpc
 import grpc
